+++ src/error.rs
@@ -5,3 +5,6 @@ pub mod dfxvm_init;
 pub mod env;
 pub mod fs;
 pub mod json;
+mod retryable;
+
+pub use retryable::Retryable;