+++ src/main.rs
@@ -17,7 +17,8 @@ mod locations;
 mod settings;
 mod style;
 
-fn main() -> ExitCode {
+#[tokio::main(flavor = "current_thread")]
+async fn main() -> ExitCode {
     let args = args_os().collect::<Vec<OsString>>();
-    cli::main(&args)
+    cli::main(&args).await
 }